requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
peewee==3.13.3
protobuf==3.13.0
PyMySQL==0.10.1
six==1.15.0
Werkzeug==1.0.1
